+++ wp-content/themes/genesis/lib/functions/upgrade.php
@@ -750,56 +750,47 @@ function genesis_silent_upgrade() {
 
 }
 
-add_action( 'genesis_upgrade', 'genesis_upgrade_redirect' );
+add_action( 'upgrader_process_complete', 'genesis_update_complete', 10, 2 );
 /**
- * Redirect the user back to the "What's New" page, refreshing the data and notifying the user that they have
- * successfully updated.
+ * Upgrade the Genesis database after an update has completed.
  *
- * @since 1.6.0
+ * After an update has been completed, send a remote GET request to `admin-ajax.php` to trigger a silent upgrade.
  *
- * @return null Return early if not an admin page.
+ * @since 2.10.0
+ *
+ * @param object $upgrader   The upgrader object.
+ * @param array  $hook_extra Details about the upgrade process.
+ * @return null
  */
-function genesis_upgrade_redirect() {
-
-	if ( ! is_admin() || ! current_user_can( 'edit_theme_options' ) || is_customize_preview() ) {
+function genesis_update_complete( $upgrader, $hook_extra ) {
+	if ( $hook_extra['action'] !== 'update' || $hook_extra['type'] !== 'theme' ) {
 		return;
 	}
 
-	genesis_admin_redirect( 'genesis-upgraded' ); // What's New page.
-
-}
-
-add_action( 'admin_notices', 'genesis_upgraded_notice' );
-/**
- * Displays the notice that the theme settings were successfully updated to the latest version.
- *
- * Currently only used for pre-release update notices.
- *
- * @since 1.2.0
- *
- * @return void Return early if not on the Theme Settings page.
- */
-function genesis_upgraded_notice() {
-
-	if ( ! genesis_is_menu_page( 'genesis' ) ) {
+	// Multiple themes are being updated but not Genesis.
+	if ( isset( $hook_extra['themes'] ) && ! in_array( 'genesis', $hook_extra['themes'] ) ) {
 		return;
 	}
-	if ( isset( $_REQUEST['upgraded'] ) && 'true' === $_REQUEST['upgraded'] ) {
-		echo '<div id="message" class="updated highlight"><p><strong>';
-		printf(
-			/* translators: 1: Genesis version, 2: URL for What's New admin page. */
-			esc_html__( 'Congratulations, you are now rocking Genesis %1$s! %2$s', 'genesis' ),
-			esc_html( genesis_get_option( 'theme_version' ) ),
-			sprintf(
-				'<a href="%s">%s</a> %s.',
-				esc_url( menu_page_url( 'genesis-upgraded', 0 ) ),
-				esc_html__( 'See what\'s new in', 'genesis' ),
-				esc_html( PARENT_THEME_BRANCH )
-			)
-		);
-		echo '</strong></p></div>';
+
+	// One theme is being updated but not Genesis.
+	if ( isset( $hook_extra['theme'] ) && 'genesis' !== $hook_extra['theme'] ) {
+		return;
 	}
 
+	$silent_upgrade_url = add_query_arg(
+		array(
+			'action' => 'genesis-silent-upgrade',
+		),
+		admin_url( 'admin-ajax.php' )
+	);
+
+	wp_remote_get(
+		$silent_upgrade_url,
+		array(
+			'timeout'  => 0.01,
+			'blocking' => false,
+		)
+	);
 }
 
 add_filter( 'update_theme_complete_actions', 'genesis_update_action_links', 10, 2 );
@@ -807,18 +798,16 @@ add_filter( 'update_theme_complete_actions', 'genesis_update_action_links', 10, 
  * Filter the action links at the end of an update.
  *
  * This function filters the action links that are presented to the user at the end of a theme update. If the theme
- * being updated is not Genesis, the filter returns the default values. Otherwise, it will provide a link to the
- * Genesis Theme Settings page, which will trigger the database upgrade.
+ * being updated is not Genesis, the filter returns the default values. Otherwise, it will provide its own links.
  *
  * @since 1.1.3
  *
  * @param array  $actions Existing array of action links.
  * @param string $theme   Theme name.
- * @return array Removes all existing action links in favour of a single link, if Genesis is
- *               the theme being updated. Otherwise, return existing action links.
+ * @return array Replace all existing action links, if Genesis is the theme being updated.
+ *               Otherwise, return existing action links.
  */
 function genesis_update_action_links( array $actions, $theme ) {
-
 	if ( 'genesis' !== $theme ) {
 		return $actions;
 	}
@@ -826,11 +815,15 @@ function genesis_update_action_links( array $actions, $theme ) {
 	return array(
 		sprintf(
 			'<a href="%s">%s</a>',
-			menu_page_url( 'genesis', 0 ),
-			esc_html__( 'Click here to complete the upgrade', 'genesis' )
+			menu_page_url( 'genesis-upgraded', 0 ),
+			esc_html__( 'Check out what\'s new', 'genesis' )
+		),
+		sprintf(
+			'<a href="%s">%s</a>',
+			admin_url( 'customize.php?autofocus[panel]=genesis' ),
+			esc_html__( 'Theme Settings', 'genesis' )
 		),
 	);
-
 }
 
 add_action( 'admin_notices', 'genesis_update_nag' );