+++ wp-content/themes/genesis/lib/init.php
@@ -85,6 +85,7 @@ function genesis_theme_support() {
 	add_theme_support( 'post-thumbnails' );
 	add_theme_support( 'title-tag' );
 	add_theme_support( 'automatic-feed-links' );
+	add_theme_support( 'body-open' );
 	add_theme_support( 'genesis-inpost-layouts' );
 	add_theme_support( 'genesis-archive-layouts' );
 	add_theme_support( 'genesis-admin-menu' );
@@ -201,10 +202,10 @@ function genesis_constants() {
 	// Define Theme Info Constants.
 	// phpcs:disable WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedConstantFound
 	define( 'PARENT_THEME_NAME', 'Genesis' );
-	define( 'PARENT_THEME_VERSION', '2.9.1' );
-	define( 'PARENT_THEME_BRANCH', '2.9' );
-	define( 'PARENT_DB_VERSION', '2900' );
-	define( 'PARENT_THEME_RELEASE_DATE', date_i18n( 'F j, Y', strtotime( '20 March 2019' ) ) );
+	define( 'PARENT_THEME_VERSION', '2.10.1' );
+	define( 'PARENT_THEME_BRANCH', '2.10' );
+	define( 'PARENT_DB_VERSION', '2901' );
+	define( 'PARENT_THEME_RELEASE_DATE', date_i18n( 'F j, Y', strtotime( '07 May 2019' ) ) );
 
 	// Define Parent and Child Directory Location and URL Constants.
 	define( 'PARENT_DIR', get_template_directory() );
@@ -349,6 +350,7 @@ function genesis_load_framework() {
 		require_once $admin_dir . 'use-child-theme.php';
 		require_once $admin_dir . 'sanitization.php';
 		require_once $admin_dir . 'privacy-requests.php';
+		require_once $admin_dir . 'plugin-install.php';
 		require_once $admin_dir . 'onboarding/theme-activation.php';
 		require_once $admin_dir . 'onboarding/ajax-functions.php';
 		require_once $functions_dir . 'onboarding.php';
@@ -374,7 +376,11 @@ function genesis_load_framework() {
 
 	// Load CLI command.
 	if ( defined( 'WP_CLI' ) && WP_CLI ) {
-		include PARENT_DIR . '/lib/classes/class-genesis-cli-command.php';
+		WP_CLI::add_command( 'genesis db', 'Genesis_Cli_Db_Command' );
+		WP_CLI::add_command( 'genesis core', 'Genesis_Cli_Core_Command' );
+		WP_CLI::add_command( 'genesis setting', 'Genesis_Cli_Settings_Command' );
+		WP_CLI::add_command( 'genesis option', 'Genesis_Cli_Settings_Command' );
+		WP_CLI::add_command( 'genesis', 'Genesis_Cli_Command' );
 	}
 
 	global $_genesis_formatting_allowedtags;